Add a `compareProfile` utility function to compare profiles
