+++ ui/packages/shared/profile/src/index.tsx
@@ -12,11 +12,7 @@
 // limitations under the License.
 
 import type {Props as CallgraphProps} from './Callgraph';
-import ProfileExplorer, {
-  DEFAULT_DASHBOARD_ITEMS,
-  compareProfile,
-  getExpressionAsAString,
-} from './ProfileExplorer';
+import ProfileExplorer, {getExpressionAsAString} from './ProfileExplorer';
 import ProfileTypeSelector from './ProfileTypeSelector';
 
 export * from './ProfileIcicleGraph/IcicleGraph';
@@ -31,10 +27,4 @@ export {default as Callgraph} from './Callgraph';
 
 export type {CallgraphProps};
 
-export {
-  ProfileExplorer,
-  ProfileTypeSelector,
-  DEFAULT_DASHBOARD_ITEMS,
-  getExpressionAsAString,
-  compareProfile,
-};
+export {ProfileExplorer, ProfileTypeSelector, getExpressionAsAString};
